Refactor setup into separate duck
