+++ src/universal/modules/meeting/ducks/index.js
@@ -1,8 +1,10 @@
 import {combineReducers} from 'redux-immutablejs';
 import meeting from './meeting';
+import setup from './setup';
 import shortcuts from './shortcuts';
 
 export default combineReducers({
   meeting,
+  setup,
   shortcuts
 });
