+++ src/universal/redux/makeReducer.js
@@ -1,13 +1,13 @@
 import {compose} from 'redux';
 import {combineReducers} from 'redux-immutablejs';
 import auth from '../modules/auth/ducks/auth';
-import meeting from '../modules/meeting/ducks/index';
+import meetingModule from '../modules/meeting/ducks/index';
 import {routing} from './routing';
 
 // TODO make auth and meeting async
 const currentReducers = {
   auth,
-  meeting,
+  meetingModule,
   routing
 };
